starting to develop Agents
RepositoryUsers will only be able to access the REST API with an agent api_token and
a user api_token.

Agents are meant to name and describe discreet computational systems that interact with
the system in a defined, repeatable, and reusable manner.  To facilitate this, they contain
links to homepages and docker_registry_urls to help users find more information about them and use them.

Model implemented and tested
  Non RepositoryUsers cannot access information about Agents
  Any RepositoryUser can index and show any Agent, create new Agents, edit/update and destroy their own Agents
  Admin RepisitoryUsers can also edit/update and destroy other user agents to enable and disable them

Users now have_many agents

controller needs to be implemented and tested